WIP: Refactoring operators
