+++ quads-query/src/main/java/fr/cnrs/liris/jpugetgil/sparqltosql/sql/context/StSOperator.java
@@ -1,0 +1,7 @@
+package fr.cnrs.liris.jpugetgil.sparqltosql.sql.context;
+
+import fr.cnrs.liris.jpugetgil.sparqltosql.SQLQuery;
+
+public abstract class StSOperator {
+    public abstract SQLQuery buildSQLQuery();
+}
